test(tests): dispose files after tests
