+++ JsonStores.Tests/Stores/JsonStoreWithExistingFile.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Threading.Tasks;
 using JsonStores.NamingStrategies;
+using JsonStores.Tests.Helpers;
 using JsonStores.Tests.Models;
 using Xunit;
 
@@ -23,9 +24,8 @@ namespace JsonStores.Tests.Stores
             {
                 NamingStrategy = new StaticNamingStrategy(_fileName)
             };
-            var filePath = Path.Combine(options.Location, $"{_fileName}.json");
-            JsonFileCreator.CreateStore(filePath);
 
+            JsonFileCreator.CreateStore(_fileName);
             _store = new JsonStore<Person>(options);
         }
 
@@ -64,7 +64,7 @@ namespace JsonStores.Tests.Stores
         {
             GC.SuppressFinalize(this);
 
-            File.Delete(_fileName);
+            File.Delete(FilePathEvaluator.GetFilePath(_fileName));
         }
     }
 }
